% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allom_calculate.R
\name{allom_calculate}
\alias{allom_calculate}
\title{Calculating new variables based on the allometries formula}
\usage{
allom_calculate(data, ..., allometry_id, name = allometry_id)
}
\arguments{
\item{data}{data frame with at least, the independent variables needed for calculating
the allometry}

\item{...}{variable names for the independent variables}

\item{allometry_id}{character with the unique allometry identifier}

\item{name}{Optional, the name of the variable created. If not provided, the name will
be the allometry_id}
}
\description{
Given the allometry id, this function creates a new variable with the desired allometry
}
\details{
This is a tidy-friendly funcion, meanining it expects a dataframe and will return
a dataframe with the new variable calculated.
}
\examples{

data_example <- data.frame(Dn = rnorm(10, 12, 2))

data_example \%>\%
  allom_calculate(
    Dn = Dn,
    allometry_id = 'BAT_1',
    name = 'BAT'
  )

}
