% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{allom_description}
\alias{allom_description}
\title{Get allometries descriptions as a list}
\usage{
allom_description(..., id = NULL)
}
\arguments{
\item{...}{Expresions resolving to a logical value to filter the allometries. Only
evaluated if \code{id} is NULL.}

\item{id}{Character vector with the allometry/ies id}
}
\value{
a list with the allometries and their
}
\description{
This function returns the allometries descriptions
}
\details{
This function can return the description of the alometries given in the id argument
or can filter the general allometries table with the expressions given in the ...
argument
}
\examples{

# by id
foo <- allom_description(id = "GC_2589")
foo$GC_2589$dependent_var
foo$GC_2589$param_a

# filtering
ht_dn_allometries <- allom_description(dependent_var \%in\% c("GC", "Dn"))
ht_dn_allometries$GC_2589$dependent_var
ht_dn_allometries$GC_2589$param_a

}
