% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{allometr_app}
\alias{allometr_app}
\title{function to launch the allometr app}
\usage{
allometr_app(user = "guest", password = "guest", host = NULL,
  port = NULL, dbname = "allometr_db")
}
\description{
function to launch the allometr app
}
