% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_data}
\alias{mod_data}
\title{mod_data server function}
\usage{
mod_data(input, output, session, allometries_table, variables_thesaurus,
  cubication_thesaurus)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{allometries_table}{table with allometries info from allometr_db}

\item{variables_thesaurus}{table with variables info from allometr_db}

\item{cubication_thesaurus}{table with cubication info from allometr_db}
}
\description{
mod_data server function
}
