% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_dataInput}
\alias{mod_dataInput}
\title{mod_dataInput and mod_data}
\usage{
mod_dataInput(id, allometries_table, variables_thesaurus,
  cubication_thesaurus)
}
\arguments{
\item{id}{shiny id}

\item{allometries_table}{table with allometries info from allometr_db}

\item{variables_thesaurus}{table with variables info from allometr_db}

\item{cubication_thesaurus}{table with cubication info from allometr_db}
}
\description{
A shiny module to create and populate the data inputs
}
