% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_equationOutput.R
\name{mod_equation}
\alias{mod_equation}
\title{mod_equation server function}
\usage{
mod_equation(input, output, session, allometries_table,
  variables_thesaurus, cubication_thesaurus, data_reactives,
  table_reactives)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{allometries_table}{table with allometries info from allometr_db}

\item{variables_thesaurus}{table with variables info from allometr_db}

\item{cubication_thesaurus}{table with cubication info from allometr_db}

\item{data_reactives}{data reactive inputs from mod_dataInput}

\item{table_reactives}{table reactive inputs from mod_tableOutput}
}
\description{
mod_equation server function
}
