% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_equationOutput.R
\name{mod_equationOutput}
\alias{mod_equationOutput}
\title{mod_equationOutput and mod_equation}
\usage{
mod_equationOutput(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny module to create and populate the equation output
}
