% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableOutput.R
\name{mod_table}
\alias{mod_table}
\title{mod_table server function}
\usage{
mod_table(input, output, session, allometries_table, variables_thesaurus,
  cubication_thesaurus, data_reactives)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{allometries_table}{table with allometries info from allometr_db}

\item{variables_thesaurus}{table with variables info from allometr_db}

\item{cubication_thesaurus}{table with cubication info from allometr_db}

\item{data_reactives}{data reactive inputs from mod_dataInput}
}
\description{
mod_table server function
}
