% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_data}
\alias{mod_data}
\title{mod_data server function}
\usage{
mod_data(input, output, session, allometr_table, variables_thesaurus,
  cubication_thesaurus, lang)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{allometr_table}{allo table}

\item{variables_thesaurus}{thes table}

\item{cubication_thesaurus}{cubication table}

\item{lang}{lang value}
}
\description{
mod_data server function
}
