% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_data}
\alias{mod_data}
\title{mod_data server function}
\usage{
mod_data(input, output, session, allometr_db)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{allometr_db}{pool object to access the allometries db}
}
\description{
mod_data server function
}
