% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataInput.R
\name{mod_dataInput}
\alias{mod_dataInput}
\title{mod_dataInput and mod_data}
\usage{
mod_dataInput(id, params, label = NULL,
  btn_label = "Restableix els filtres", inline = TRUE)
}
\arguments{
\item{id}{shiny id}

\item{params}{list with params}

\item{label}{label}

\item{btn_label}{button label}

\item{inline}{logical}
}
\description{
A shiny module to create and populate the data inputs, linked between them.
  It is based on shinyWidgets::selectizeGroupUI module, but adapted to our case
}
